Enable HPSS archiving on C5/6 (#3437)
# Description
This PR enables HPSS archiving on Gaea C6. The C5 hosts file is also
updated, but has not been tested. The ES cluster needs to know which
nodes to select based on which file system the files to archive are
located. To accomplish this, the `--constraint=` option was added for C5
and C6. This feature should have no change on other platforms.

Resolves #3421 
Refs #3261 

# Type of change
- [x] New feature (adds functionality)

# Change characteristics
- Is this a breaking change (a change in existing functionality)? NO
- Does this change require a documentation update? YES/NO
- Does this change require an update to any of the following submodules?
NO

# How has this been tested?
C48_ATM test on Gaea C6

# Checklist
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have documented my code, including function, input, and output
descriptions
- [x] My changes generate no new warnings
- [x] New and existing tests pass with my changes
- [x] This change is covered by an existing CI test or a new one has
been added
- [x] I have made corresponding changes to the system documentation if
necessary