Improve messaging to display clear warning when missing snogrb file (#3317)
Resolve bugzilla 1373 (Display clear warning when missing snogrb file). 
The following changes are included in this PR to resolve the bugzilla:

- add "INFO" message when final snogrb file is decided in script
- add "WARNING" message when decided-upon snogrb file is missing and state that "Snow coverage will not be updated."

Also included in this PR are a related bug fix uncovered during work to
resolve bugzilla and needed COMIN/COMOUT updates, which are part of a
larger initiative. Those changes are:

- fix bug with `res` variable setting in `scripts/exgdas_enkf_sfc.sh`
- COMIN/COMOUT updates for the sfcanl jobs

Also included in this PR is the removal of `getges.sh` from the related
scripts and removal of `GETGES` mention in some wave job scripts. Full
removal of `getges.sh` will be done in a separate PR.

Refs #2451
Refs #1041
Resolves #2500