Updates for Gaea: verif-global tag, tracker tag, Fit2Obs tag, and C768 analysis resources (#3463)
# Description
This PR updates the tags for verif-global, the TC tracker, and Fit2Obs
to include Gaea support for each. Additionally, it updates the resources
for the C768 and C1152 analyses to use the same resources as WCOSS2.
This allowed the C768 analysis to run to completion for a few test dates
(2024092600-2024092618). C1152 has not been tested. Lastly, it reenables
support for METplus jobs on Hera.

Resolves #2639 
Resolves #3402 
Resolves https://github.com/NOAA-EMC/global-workflow/issues/3449
Resolves https://github.com/NOAA-EMC/global-workflow/issues/3370
Resolves #3371

# Type of change
- [x] Bug fix (fixes something broken)
- [x] New feature (adds functionality)
- [x] Maintenance (code refactor, clean-up, new CI test, etc.)

# Change characteristics
- Is this a breaking change (a change in existing functionality)? NO
- Does this change require a documentation update? YES
- Does this change require an update to any of the following submodules?
YES
- [x] EMC verif-global
https://github.com/NOAA-EMC/EMC_verif-global/pull/138

# How has this been tested?
The tracker has been tested on all platforms. Fit2Obs has been
extensively tested on C6, but not elsewhere. The analysis resources were
tested on C6 as well for a cycled hybrid 4DVar case. Hera support for
verif-global was also tested.

Additional testing will be performed through CI tests on WCOSS2 before
marking this ready for review.

# Checklist
- [x] Any dependent changes have been merged and published
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have documented my code, including function, input, and output
descriptions
- [x] My changes generate no new warnings
- [ ] New and existing tests pass with my changes
- [x] This change is covered by an existing CI test or a new one has
been added
- [x] I have made corresponding changes to the system documentation if
necessary

---------

Co-authored-by: kate.friedman <kate.friedman@noaa.gov>