Update AWS defaults for running obs prep jobs on cloud (#3681)
This PR enables running prep jobs on AWS. This required supporting
`obsproc` and `prepobs` on the cloud
(https://github.com/NOAA-EMC/obsproc/pull/99 and
https://github.com/NOAA-EMC/prepobs/pull/43) and deciding on placement
of `BASE_GIT` directory which houses `obsproc` and `prepobs`. `BASE_GIT`
is proposed to be `/contrib/git` on AWS (and other NOAA CSPs), and
`obsproc` and `prepobs` have been cloned and installed in `/contrib/git`
on the EPIC's AWS `/contrib` filesystem for future CI testing. The obs
data used was made available on AWS via #3494, and the `DMPDIR` is
defaulted to `/bucket/dump` or user-defined via `DATA_MOUNT` for AWS.
Since restricted data is not yet supported on NOAA CSPs, AWS defaults
are updated to include `CHGRP_RSTPROD='NO'` and `MAKE_NSSTBUFR='YES'`.
Updating `CHGRP_RSTPROD` is necessary since NSST buffer files are
assumed in `obsproc` to consist of restricted data and are overwritten
as empty when `chgrp rstprod` fails. Allowing making the NSST buffer
file on the fly with `CHGRP_RSTPROD='NO'` was done via
https://github.com/NOAA-EMC/obsproc/pull/100.

Resolves #3497