Updates to resources and yaml files from GFS C1152 cycled testing (#3634)
Add yaml files that are being used for GFSv17 C1152 cycled testing.

Update resources that have been used to get C1152 cycled tests
running on wcoss2 and on Gaea (Note for Gaea an additional change to
config.ufs is needed to increase the number of tasks on the write grid
component. It's my understanding that because of calculations done other
places we can't put an "if gaea" there and it has to be manually
changed, so a note is added).

Refs #3227