Add AWS runtime configs for C96_atm3DVar jobs (#3792)
This PR adds resource and runtime environment configurations for running
the `C96_atm3DVar` test on AWS. Additional steps that run on the
"compute" partition are `anal`, `analdiag`, `analcalc`, and `upp` in
order to reduce the number of nodes requested in order to reduce costs.
The `anal` step has a significant change from tier 1 machines where
`threads_per_task` is between 5 and 8. Here, `threads_per_task` is set
to 1. This allows only 2 nodes be requested with their CPU usage
reaching ~90%. For `analcalc`, the `srun` option `--mpi=pmi2` needs to
but cannot be passed to the runtime `srun` command within the `analcalc`
step so that the environment variable `SLURM_MPI_TYPE` is instead
exported. Lastly, `perl` is added to `module_base.noaacloud` for
`vminmon` steps (available on shared NOAA apps snaphsots).

Resolves #3685