Bugfix for soil Increments archiving and staging (#3902)
This fixes two related bugs in archiving soil DA increments and
staging them for warm starts. The previous version was erroneously using
"DOIAU" instead of "DO_Land_IAU" to determine the GSI soil increment
files to be archived. Code to stage the increment files when needed
(cycled warm start) is also added.

A warm-start soil DA CI test is also added here as none of the existing
CI tests had a warm start (for land) and that contributed to the bug
addressed in this PR not being caught earlier.

Resolves #3900 
Resolves #3904