Specify <cores> when running on a single node without threading; enable GDASApp to build on Gaea eslogin nodes (#3989)
This modifies the way rocoto XMLs are generated so that single-node jobs
that are not threaded will use the `<cores>` resource syntax instead of
`<nodes>`. On slurm systems, this translates into submitting a job with
`--ntasks=X` instead of `--tasks-per-node=X --nodes=1-1`, which loosens
the allocation of such jobs and gives them a higher priority. This is
needed for the GDASApp builds, which require substantial memory.

Fixes #3932