HOTFIX: Disable metp jobs on MSU and C6 (#4058)
C6, Hercules, and Orion currently do not support EMC_verif-global due to
the unavailability of MET/METplus v9.1.3/v3.1.1, which are required by
the package. This is a short-term stopgap PR to disable those jobs from
running on MSU and C6 while a fix is put in place.

This also removes `DO_*` options from host yaml files that are
ignored/overridden.

Refs #4001