Update wave jobs to create filenames consistent with EE2. (#4046)
This is in preparation for GFSv17 code handoff and
 - renames wave output to be consistent with the EE2 filename convention
of `<RUN>.t<cyc>z.<domain>.<extension>`
- removes the `.wave` from the filename as it resides under the `wave/`
directory structure. This change was discussed with NCO
- cleans up affected scripts by passing them through `shellcheck` and
`shfmt` (where appropriate).

This also overrides the output in `ocean.stats` file via `MOM_override`.
This was brought up in #3931.  
After discussing with @sanAkel, the above issue should be closed, since
the channels are used as fix files, override has nothing in it (of
value), and layout is the default.