Configure resources correctly for marineanalecen (#4048)
Configures resources for marineanalecen. 
With parallel recentering, there is a constraint on how many MPI tasks
can be used in the application.
Ensemble members are processed in parallel. `nens` and `nens per MPI
task` determine how many subensembles are processed in parallel. In this
PR we're set to run with 5 nens per MPI task if `nens` is divisible by
5, 2 if `nens` is divisible by 2, and 1 otherwise.
The number of MPI tasks for marineanlecen needs to be a multiple of the
number of subensembles, so `config.resources` is updated accordingly.
This is a memory-heavy application, so we're also using fewer tasks per
node.

Resolves https://github.com/NOAA-EMC/global-workflow/issues/4041