Quick update to README.md badges layout & instructions file (#4075)
This is a quick update to the layout of the badges for the Nightly CI/CD
Pipelines.
We needed to remove the legacy GitHub Action based badges that where
simply keyed of of the CI/CD label states.

Also note that we have include an update to [GitHub Copilot's
Instructions
file](https://docs.github.com/en/copilot/how-tos/configure-custom-instructions/add-repository-instructions):

This update represents enhancements from the Foundation Model's
perspective along with an assist from a MCP/RAG framework in support of
agentic base AI software development. This instruction file will be
inspected by any action on the repo by GitHub Co-pilot and give it a
lean in on what to look for and how best conduct its work.