Write atmospheric history files at a different resolution than the model (#4030)
This adds the ability to run the model at one atmospheric horizontal
resolution and write the history file and master grib file output at a
different resolution. The resolutions supported by default are the
target operational atmospheric resolutions (C1152/C768) and the CI test
resolution (C96/C48). Other resolution combinations would require
additional FIXWGT files to be generated by the user (such as in #3937).

By default, the model forecast resolution and the output resolutions are
set to be the same (CASE=$CASE_HIST). The CI test
C96C48mx500_S2SW_cyc_gfs was updated to exercise this capability.

The benefit of this feature is to reduce disk space for the output while
still maintaining the skill gained from a higher resolution forecast.
Since the atmospheric DA is already run at a coarser resolution than the
model forecast, there is not a negative impact to skill when reducing
the history file resolution. See results here:
https://docs.google.com/presentation/d/1TAqe-nOFfD4odDKbyY4bZc5klXxlKFHvkYpnEK5KL8M

Resolves #3619