Add attributes to Gaussian grid sfcanl file (#4149)
The gaussian grid sfcanl file does not have the needed global attributes
for upp. The attributes of imp_physics and landsfcmdl are missing and
needed for upp to compute cloud and land fields consistently with the
model. Upp assumes default values but the imp_physics default is
inconsistent with what is used in GFSv17. This PR adds the needed
attributes to the script and namelist for the gaussian sfcanl utility
and updates the gfs_utils hash that allows the utility to read the new
namelist.

Resolves https://github.com/NOAA-EMC/global-workflow/issues/4113
Refs: https://github.com/NOAA-EMC/GFS/issues/29
https://github.com/NOAA-EMC/gfs-utils/issues/147