Update bash ex-scripts to be linter compliant (#4151)
Updates the ex-scripts to be compliant with both `shellcheck` and
`shfmt`. This does _not_ necessarily enforce some standards not checked
by `shellcheck`. Export statements are added to some variables to avoid
'not used' errors. In the future, a more careful analysis should be done
to see if these variables are used downstream by executables or if they
are truly unused, in which case they should be removed.

Scripts employing MPMD are also updated to use `run_mpmd.sh`. This
resulted in some retooling of those scripts.