Update shell script analysis GH action (#4280)
Updates the GitHub action for static shell code analysis. The shellcheck
is switched over to the reviewdog ecosystem and now checks all files in
directories that have been made compliant, rather than just checking
changes. A second pass on just changes is temporarily included (also
using reviewdog, as that mode is supported) to ensure any new changes
are compliant until the entire workflow is brough up to spec.

There is also a new check against `shfmt`. The reviewdog version does
not allow for path restrictions for shfmt, so for now this is being done
manually on compliant directories. The reviewdog version is included to
be uncommented when the entire repo is compliant.

Additionally, the bash and python GH workflows are renamed to better
represent their contents and updated to have the same set of triggers.
The push trigger is restricted to develop to avoid double-testing of
PRs, and a workflow dispatch trigger is added.

Finally, a couple scripts that became non-compliant due to a lack of a
shfmt check have been corrected.

Refs: #397
Resolves #4239