Fix eobs diags being left on the floor and a number of EE2 related updates (#4267)
- Tar up the gsidiags into a
gsidiag.tar instead of the GSI writing out to COM via symlinks. The
tarred GSI diags are then untarred where they are needed.
- provides a number of EE2 related updates (replacing link with copy),
`COMIN/COMOUT`, etc.

During the course of this , it was noticed that the EnKF was not
assimilating any satellite obs. This hotfix was issued by
@CatherineThomas-NOAA in a different PR already.
Several variables in the analysis-related scripts are unnecessary and
seem to come from earlier copy/paste versions. They can lead to
confusion. This PR tries to eliminate some, though more work is
recommended before code hand-off to expedite code acceptance into ops.
There are also several deprecated "options" that are either no longer
functional or not used.