Update remaining bash scripts to be compliant with static bash code analysis checkers (#4329)
Updates all the remaining bash scripts to be complaint with the static
bash code analysis checkers. With this update, all bash scripts with the
global-workflow repo should be compliant with `shfmt` and `shellcheck`.

The `setup_ecf_links.sh` scripts was the only one that required
substantive refactoring. The existing one didn't work as expected. The
refactoring both made the script compliant and restored the correct
functionality. One section linking the awips g2 directory has been
temporarily commented out as that directory has not yet been copied from
v16 to develop.

The manual documentation builder script was renamed to have a `.sh`
extension.

Now that the entire workflow is compliant, the GH action removes the
temporary actions and switches to using just the reviewdog actions on
the full repository.