Revamp obs staging and analysis stats job
This makes changes in two areas of GW code for JEDI jobs.

First, it makes changes to `ush/python/pygfs/jedi/jedi.py` related to
observations handling that does the following:
1. It creates a `stage_observations()` method for the `Jedi` class that
stages observations for analysis jobs, rather than relying on the task
config YAML in GDASApp for staging. This change is justified by the fact
that obs staging is essentially the same across all analysis tasks.
2. Methods for staging, extracting, taring bias corrections, and saving
obs diags and radiative bias corrections are moved from the the
`Analysis` class into the `Jedi` class for the following reason:
3. The paths, file prefixes, and file suffixes used in these methods for
obs, obs diags, and bias corrections are taken from the JCB config
dictionary in the `Jedi` class. This ensures that the file structure and
naming for obs and their statistics are consistent between how they are
staged in GW and how they are stages and saved by JEDI applications.
Thus there can never be naming conflicts.
4. The JCB config dictionary for a `Jedi` object is created by the class
constructor rather than the `initialize` class. This way, `task_config`
doesn't need to be passed to both the class constructor and the
`initialize` class. One benefit of this is that it cuts down on the
number of times `task_config` is dumped by the logger.
5. Other minor changes are made to the `Jedi` class code such as
hardening and more descriptive method/variable naming.

Second, `ush/python/pygfs/task/analysis_stats.py` is refactored in the
following ways:
1. `AnalysisStats` class now inherits from `Analysis` rather than
`Task`. This allows it to inherit parameters like `APREFIX`, `GPREFIX`,
etc.
2. Changes are made so that `task_config` is never modified after the
class constructor is run, consistent now with all other tasks.
3. The "base config" and "JEDI config" YAMLs are consolidated into a
single master config YAML (like all other task now), and any
staging/saving that was carried on in the original Python code is now
invoked by the FileHandler with that master YAML (`data_in` and
`data_out` keys).
4. The run directory is reorganized by analysis type with subdirectories
for the inputs and outputs.
5. Input and output paths for staging/saving are taken from the JCB
config dictionary of the relevant `Jedi` object, to ensure that file
paths and naming are consistent between the GW code and JEDI application
configuration YAMLs.