Fix bash code analysis issues
Fixes some issues from the recent expansion of bash code analysis and
switch to the reviewdog framework.

First and foremost, the permission issue that prevented errors from
being posted to the PR are fixed. This was only an issue for pull
requests from forks (which is all of them for us). The trigger is
changed to `pull_request_target`, which changes the context to the
target repo. This is safe as long as we are not attempting to build or
run the code. Because of the context change, an additional argument is
needed for checkout to checkout the head of the PR.

Second, some parameters are turned into conditionals to handle the fact
that the action is run on both pull requests and merges to develop. This
allows both the correct hash in the previous change, and also the
correct GH reporter to be used. This uses a new GH action that will need
to be added to the allowed list.

Third, the `shfmt` options are fixed so that it actually fails on error.
Reviewdog has deprecated `fail_on_error`, but the `reviewdog/shfmt`
action does not seem to support the replacement `fail_level`. So,
`fail_level` is passed to reviewdog directly as a flag.

Finally, the README status badges are updated for the new actions.
Additionally, clicking on the code analysis badges now brings up the GH
action logs.

Note: the `bash_code_analysis` actions may not run for this PR because
of the trigger change. It will run on develop after merge, and on PRs
from then on.